xxx/requirements.txt
numpy
pandas
statsmodels
matplotlib
python-dotenv